move ScreenBrightness handler to thermostat module
