+++ tests/entities/number.py
@@ -7,9 +7,7 @@ from custom_components.meross_lan.devices.garageDoor import (
 )
 from custom_components.meross_lan.devices.hub import MLHubSensorAdjustNumber
 from custom_components.meross_lan.devices.mts100 import Mts100AdjustNumber
-from custom_components.meross_lan.devices.screenbrightness import (
-    MLScreenBrightnessNumber,
-)
+from custom_components.meross_lan.devices.thermostat import MLScreenBrightnessNumber
 from custom_components.meross_lan.merossclient import const as mc
 from custom_components.meross_lan.number import MLConfigNumber
